Componentized the Blog Page.
