+++ apps/web/src/components/blog-category-nav.tsx
@@ -1,5 +1,4 @@
-"use client";
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Link from "next/link";
 
 interface Category {
