+++ apps/web/src/lib/sanity/query.ts
@@ -249,6 +249,8 @@ export const queryBlogPaths = defineQuery(`
   *[_type == "blog" && defined(slug.current)].slug.current
 `);
 
+// Blog category page data
+
 export const queryAllCategories = defineQuery(`
   *[_type == "category"] | order(title asc) {
     _id,
@@ -263,9 +265,13 @@ export const queryAllCategories = defineQuery(`
   }
 `);
 
+// To get all category slugs
+
 export const queryAllCategorySlugs = defineQuery(`
   *[_type == "category" && defined(slug.current)]{ "slug": slug.current }
 `);
+
+// To get all blogs by category slug
 
 export const queryBlogsByCategorySlug = defineQuery(`
   *[_type == "blog" && references(*[_type == 'category' && slug.current == $slug][0]._id)] | order(orderRank asc) {
@@ -284,6 +290,8 @@ export const queryBlogsByCategorySlug = defineQuery(`
     ${pokemonFragment}
   }
 `);
+
+// To get a single category by slug
 
 export const queryCategoryBySlug = defineQuery(`
   *[_type == "category" && slug.current == $slug][0]{
